Hacky attempt at Babel-ifying in the browser
Committing this for posterity, you can see a couple of different hacky
things I did trying to get this to work.

I'll back this out once I figure out the right way to do it. I suspect
that the most appropriate choice will be a build process that dumps
something in a "dist" directory. At some point I'll want to turn on
"production" mode as well so that the `__DEV__` checks in React get
skipped over (or ideally even stripped out).