Add scratch.html
This is a playground for me to experiment with layout ideas before I
incur the overhead of building everything out in terms of
well-abstracted components.

One of the neat things about Electron is I am coding for a single,
modern browser, so I can use things like flex-box without vendor
prefixes or worrying about fall-back support.