Improve appearance
- add gray background
- remove gaudy debugging borders
- style separator (gradient, edges)
- add bottom border to `OmniBar`

I also noticed on adding the background color that it doesn't extend
over the entire {left,right} subpane area. This is partly cured (for
width) by making each pane `display: flex` and the children `flex-grow:
1`, but it only handles the horizontal direction; I am not sure how to
simultaneously handle the vertical direction.

As a temporary(?) workaround I added the background color on the
`SplitView` itself, noting the smell in a comment.