Truncate note text in NotePreviews
This makes things much snappier. Instead of storing the potentially huge
note text in the DOM (for all notes), we store 250 characters. This is
enough to always at least fill two lines for the preview at a width of
600px.

I added this as a maximum width to match the behavior of nvALT. Note our
constraint handling still isn't perfect, because it only applies when
dragging the separator: if you resize the window, we let flexbox take
over, and at that point it can violate our constraints. I'll probably
add a resize listener to handle this.