Improve split dragging and resizing behavior
Add a window resize listener so that we can adjust the split on window
resize. Note that if I end up adding more resize listeners I'll abstract
this into a module so that we can have a single listener that is shared
by others.

Additionally, we tighten up the behavior to match nvALT: dragging to the
left clamps for a while at 75px, then clamps to 0px.

So, now we basically preserve our constraints even when resizing the
window. Nice.

Known bug:

- when you resize the window in a big way, we get a subtle drifting of
  the separator, which I attribute to rounding errors in flexbox caused
  by the space taken up by the separator

Despite this, this is much more solid than what we had before. I feel
pretty good about it.