Fix grabbing cursor inconsistencies
This (setting a `cursor` style on the `<body>` whenever dragging) used
to work before), but stopped when we set an inline `cursor: default`
style on an element. Note that we could override `cursor: inherit` using
this technique, but then we won't get the desired `cursor: default` any
more during normal hovers (we'd get a text-selection cursor).

Only an `!important` style can override the `cursor: default`, and you
can't set those like this:

  document.body.style.cursor = 'blah !important'

So, we need a class:

  * { cursor: blah !important; }

Would be enough, but we also need a way to toggle it conditionally, so
we set it as with a class on the body. Yes, this selector is expensive,
and I feel terrible for using it. I could town it down to `.grabbing *`
but I don't think that would actually help much.