Handle errors more robustly when reading files from disk
We're swallowing errors, so we could wind up with null entries during
our `join`. Be defensive about those by adding `&&` guards. In this
case, you'd wind up with partially empty note records, but I think they
should be valid (enough for the rest of the program not to crash, at
least).

If an error does slip through, `catch` it and give the user the choice
to quit the app.