Add clamp module/tests and set up Jest
So I should have included the `clamp` module in a prior commit but
somehow I didn't see it. So, here it is, along with tests and set-up for
Jest.

Officially, Jest doesn't work with Node 0.12, which is what I am using,
but at least this tiny test doesn't have any problems with it.

You can run with:

  npm test
  npm run test
  npm run jest
  jest # because I must have done a global install some time?

Of these, I prefer the latter.