Implement "Select All" for NoteList
As noted inline, this was way more complicated than I would like it to
be.

The trouble is that the "Select All" menu wants to send a message to the
first responder, and there's no convenient way to intercept that in
JavaScript. I _can_ figure out when the `<NoteList>` has focus, however,
and I _can_ subscribe to "selectionchanged" events.

But, it's more complicated than that. By the time the event has
happened, the selection has already occurred; you can't prevent it or
stop propagation because it is a retrospective event. You _can_ remove
the selection (ie. with `getSelection().removeAllRanges()`) but then you
see a selection flicker in and out (in the `<Note>` pane).

It's not even clear to me why `<Note>` is being selected when it doesn't
have focus here; it's evidently up to the browser whatever it decides to
do.

So, I implement a hack: we listen for "selectionchanged" only when
`<NoteList>` has focus, we make `<Note>` have `user-select: none` so
that it doesn't get selected on "Select All" (unless it has focus, in
which case it has `user-select: inherit`). The hackiest bit of all is
the `setTimeout`; as noted inline we have to delay setting up the
listener or we wind up getting "selectionchanged" whenever we focus the
`<NoteList>`.

Along the way, fixed some tabindex issues: they are supposed to be
positive, so I renumbered them (from 0 to 2 -> 1 to 3). I also moved the
tabindex onto `<Note>` instead of `<NoteView>`.

For some reason, I was seeing the `<Separator>` get focus during
development, so I made it have a tabindex of -1. Hopefully it is ok now.

One more mystery to solve: my tab order is effectively now:

- initial -> `<body>`
- 1 -> `<OmniBar>`
- 2 -> `<NoteList>`
- 3 -> `<Note>`

I don't want `<body>` ever to be the `document.activeElement`, but it
is, meaning that we have a four-tab cycle instead of the desired
three-tab one.

Alternative approaches not taken: have a "focus store" that effectively
manages first responder for us (ugh); listen to "Select All" via "ipc"
everywhere, and have everywhere implement its own focus tracking (ugh).