Get massive speed boost with `shouldComponentUpdate`
Adding `shouldComponentUpdate` to `NoteList` doesn't help much, but I
don't think it costs us much either, so I am adding it (I don't want to
rerender everything just because the focus toggled from the `Note` to
the `OmniBar`, for instance).

But adding it to `NotePreview` makes a huge difference. I was too lazy
to profile this, but the perceived responsiveness (tested by holding
down the DOWN key) is an order of magnitude better.