Improve perceived startup performance
Prepopulate the DOM with something based on what React will render. This
shows up pretty much immediately.

I got the markup by copying the actually rendered markup out of the dev
tools, then removed the contents of the div containing the lower panes
and tidied up the output. The only thing that I needed to modify was
hoisting the `#bebebe` background color up to that lower containing div.

I'll update this again later, before the first "real" release of the
app.