Show icons in OmniBar
These aren't dynamic yet; I'll do that next.

I thought there should be a hover on the cancel icon, but I see that
Apple apps like System Preferences don't have a hover (and others like
Safari don't even have a cancel icon).

nvALT has a puzzling "back" hover on the magnifying glass icon. It is
supposedly to go "back" to search (ie. clear the input field). I don't
see any point in this, and hadn't noticed it in years of using
Notational Velocity/nvALT, so I'm not going to add it.