Start thinking about note re-ordering
I originally started down a (misguided?) path here of adding a map to
the `NotesStore` from `noteID` -> note. So basically I would have a
`List` containing the (mutable) ordering of the notes within the
`NoteList`, and a mapping that would enabled me to look up a specific
note given its ID without having to scan the `List`.

    // There may be an easier way to do this, but...
    //
    // Convert: List[Map{id: b, ...}, Map{id: a, ...}, ...]¶
    //      to: Map{b: Map{id: b, ...}, a: Map{id: a, ...}, ...}
    noteMap = notes
      .groupBy(note => note.get('id'))
      .map(list => list.first());

I stopped because I realized that, if I started passing around `noteID`s
instead of `index` (within the `notes` `List`), that I would need a way
of finding the position of the note identified by `noteID` within the
list, and then I'm back in linear scan territory again. If I have a map
from `noteID` to position with in the `List`, then I have a map with an
expensive-to-update structure (because everytime you bump something from
position N up to position 0, you have to renumber everything between 1
and N).

So, let's take a step back and see if we can do something more minimal.

- Stick with the single `List` for now
- Take advantage of the fact that we have immutable data objects and
  pass full note objects down into the `NotePreview`; we don't have to
  worry about them getting mutated under us
- Use `noteID`s as React keys; this is The Right Thing(tm) to do.

As a bonus, this improves the cohesiveness of the `NotePreview`
component. It is now able to do some of the work (like title truncation)
that was previously pretty extraneous when done in its parent, the
`NoteList`.

I'll keep thinking about how to handle note re-ordering in future
commits.