Make a Store base class and use it
This is DRYer and allows us to drop a bit of boilerplate from the store
subclasses.

The one "meh" thing here is that I had to move the `change` method onto
the class instance in the `NotesSelectionStore` because it requires
access to `this.emit`, which is unfortunate because it removes it from
among the other internal myriad methods in the store.