Mostly fix problem with excess NotePreview listeners
Other than the bug noted inline in the comments, this resolves the
problem with having too many NotePreview listeners. I'll think of a
solution to solve the remaining bug, and if I can't do it internally to
the NotePreview module I'll just have to move the listening up one level
to the NoteList.