Hack around broken selection behavior
This isn't the final state I want to be in, but it shows what I've tried
so far.

This is crazy, but anything I do in the "contextmenu" handler is
deferred (eg. logging) until _after_ the menu is shown and hidden. All
events appear to be suspended while the menu is open (fair enough, I
guess).

So, the only way to get the NotePreview selection to show up before
showing the menu is to put the menu in a timeout. Ugh. 100ms was
empirically determined to be near the lower threshold of what will work,
at least on this machine...