Move contextual menu code into main process
I took a quick look at Atom and it does all of its contextual menu stuff
in the main process, so I may as well do the same.

I was hoping this would address my lag issues, but I'm afraid it
doesn't, so another solution for that will be required.