Make @pure decorator collaborate with component method
I was thinking I might need this for some `forceUpdate` shenannigans
(specifically, trying to get the `NotePreview` to focus when
right-clicked, before the contextual menu shows up), but that was
starting to get real messy, and I'm skeptical that it would actually
work in the long run (as the menu lag issues I am fighting against seem
pretty deeply grained).

Nevertheless, I see this as being pretty useful in a Flux-driven
application for the reasons noted in the docs. Let's ship it!