Creating almost-release version of app and use it if available
With this, you'll get the Corpus app with the correct name and icon if
you've previously run `gulp release`.

The actual way I'm getting gulp to copy and rename files is pretty
clowny, but before I go looking for the ultimate pattern I just wanted
to get the basic proof of concept working.