Speed up Command-{J,K} for Note view too
The interplay here between the Note view and performKeyboardNavigation
is going to get more complicated over time, so I'll want to keep an eye
on it, but my intuition is that this is really the only place where
there will be this kind of coupling (the other bits of the app can just
call `performKeyboardNavigation` without much special casing), so it
might be ok.

If it becomes a problem, I'll think of a pattern to handle it.