Fix whitespace issue on window resize
This fixes the unwanted visible white space on the right side of the
window when resizing.

Testing showed that this was happening when I had a lot of content in
the `NoteList`. I could make it go away by reducing the note count or
reducing the amount of text in each note.

So this commit takes the "fixed data table" approach, rendering only the
visible rows and a few on either side of the viewport.

It works pretty well, even without throttling, although note the
weirdness I report when reducing the buffer on either side of the
viewport (we lose momentum scrolling).