Use simpler DOM in NoteList
Trying to see if I can make layout cheaper for the browser by using
absolute positioning instead of explicit spacer elements.

Is it faster? Not perceptibly so, but it is conceptually cleaner at
least.

The throttling may need to be tweaked as things still feel a bit janky
(momentum scrolling feels off).