Tweak throttle numbers
So, our momentum scrolling problem is due to the expense of the work
we're doing. These numbers mean we update less frequently and momentum
is mostly better, but it's still not perfect. Also, you can visually see
the blank space when you flick fast. This may be ok because it is not a
common use case (the main UX flow is finding by filtering, not
scrolling).

I'd still like to make this faster though, so I will probably loop back
in the future to try and make things better.