Use requestAnimationFrame
I don't know if this makes it any better; I thought it might on my first
try (was able to scroll from the bottom all the way to the top of 582
notes which 4 flicks), but then on trying again it was about the same as
without `requestAnimationFrame` (5 flicks).

In any case, this seems The Right Thing to do, I guess. The only real
way to make this much faster will be to optimize the work that I'm
asking React to do (do less of it, or make the work that it has to do
easier).