Proof-of-concept to demo search speed
Once you've typed a word, speed is about 3-5ms per keystroke. Earlier on
(ie. if you just type "x") you see slower results (80-500ms depending on
the letter and therefore the number of results returned). This might be
due to my logging, as when I take that out I see the time go down to 1
or 2ms, so take all this with a grain of salt.

We'll probably still want debouncing, and might want to do a special
hack to optimize single letter searches. We'll deal with that when we
come to it.