Fix potential latent bugs in NoteSelectionStore
Our `change` method relies on being able to compare the value of
`selection` before and after effecting a change operation. We were
mutating `selection` during some of these operations, instead of
returning it in an idempotent way.

No real bug here because we were saving the previous value in a
temporary variable, but it's still safer to make sure all these change
functions are pure.