Start teasing apart selection handling code
I think I'll be able to make this work much more effectively if I
separate the effects of subscribing to the different stores. For typical
updates, we were seeing two calls to `_updateNote`: once when we updated
the `FilteredNotesStore`, and then the `NotesSelectionStore`. In reality
we didn't want to update the UI until both of these had finished
happening.

With this change, we still might update twice, but at least we'll be
able to separate the effects.