Take a working snapshot before making big changes
I just noticed I was wrong about nvALT's selection behavior in the case
of finding an exact title prefix match. It doesn't actually bump; it
just selects. (It looks like it bumps only because the more you type,
the more likely you are to see the title match at the top.)