Improve focus handling
This is still a work in progress. In this commit we make it so that
tabbing to/from the Note.react component works as you'd expect. We do
this by intercepting the `<Tab>` key and handling it manually. I haven't
fully followed through with doing the same for the `OmniBar` because I
want to better match nvALT's behavior here (it doesn't have a crude
cycle; it has some special semantics depending on the state of the
application).