Split textarea out into a ContentEditable component
The `ContentEditable` name is wishful thinking for now; this is still
just a textarea, but I wanted to move it out of `Note` in order to
reduce the number of concerns it had to deal with. This is useful
because as I add contenteditable functionality, things may start to get
complicated.

Despite the name, this is unfortunately not a very "pure" container for
editable content (it takes a `note` prop, for instance, and it knows
about stores); we'll see if I can gradually beat it into shape as I
continue to work on it.