Use `npm shrinkwrap`
For one thing, this is going to be more robust, as we're locking all our
versions down.

For another, I wanted to do this to get the lastest version of
babel-eslint, which fixes a bug with decorators (specifically, they
weren't being visited, so they got flagged as unused variables).

In order to update packages from here on `npm install some-package`,
then test, then `npm shrinkwrap` again. Or at least, that's what I think
so from reading:

  https://docs.npmjs.com/cli/shrinkwrap