Add "lint" to "watch" task
Note we don't fail or ring the bell here if something goes wrong (and I
do want to do that), but this basically works, even without use of the
`wrap` function.

Note that I am running the JS transpilation and the linting in parallel,
which is probably a stupid thing to do.