Implement delete
This mostly works, but I'm getting an invariant about dispatching during
a dispatch which I need to track down. Despite the invariant, things are
working from the user perspective, so I probably just need to find the
right place to add a call to `if (!Dispatcher.isDispatching()) ...` and
add it.

Note that this required somewhat of a reworking of how the
`FilteredNotesStore` and friends work. The `NotesSelectionStore` manages
a set of indices into the `FilteredNotesStore`, but in order to delete,
we need indices into the `NotesStore` (because the dependency is
`NotesSelectionStore` -> `FilteredNotesStore` -> `NotesStore`, and in
order to avoid circular dependencies in our imports, we need to be able
to pass the `NotesStore` ids into the action handler, modify
`NotesStore` first, and then have the change trickle down to the other
stores).

Rather than building a parallel data structure to map the filtered IDs
to the canonical IDs, it seemed better to just decorate the filtered
notes with the additional index information. This is still a fair bit of
work, but perceived perf still seems to be ok despite the extra object
allocations.