Show confirmtion dialog before deleting
Unlike nvALT, I don't think we can show subtext in the drop-down
sheet ("Press ⌘-Z to undo this action later."), nor can we show
the checkbox with "Do not show this message again", nor can we
customize the primary button text (ie. it shows "OK" instead of
"Delete").

But them's the breaks, unless we want to invent a sheet replacement
which will be a bit of work and will never look native (unfortunately
we're going to have to do exactly this for the prefs).