Real filesystem writes for note creation
This is scary. Mutating the filesystem for realz now.

On the bright side, this fixed the bug I was seeing with note focus not
happening. A little bit of asynchrony goes a long way.

Note the use of fsync; not sure if this is overkill, but given that I am
using Git to store history here, I want to make sure it will be able to
"see" any changes that happen.