Fix lint error for OperationsQueue
This fixes a use-before-define warning, but I am not sure I am very
happy with the result. The thing wasn't really "used" before it was
defined. It merely appeared in the source file higher up, but doesn't
end up getting used until much later.

Doing this for now to shut-up the linter, but I am strongly tempted to
turn off this lint.