Implement real deletion, in a way
Doesn't fully work (somehow I've busted in-memory deletion as well), but
this is a rough sketch of what will be required. As noted in the inline
`TODO`, we'll need to come up with a way to force-flush edits of the
actively-edited note (at high, pri) in order for this pre-delete commit
to actually be worth anything.