Munge illegal chars in paths
"/" is the only illegal character, so change it to "-" at note creation
time (or rename time).

This means the next time the app is launched, we re-read the note and
consider any hyphens in the title to be titles, so this is a lossy
conversion. Maybe later we can consider storing the true title out of
band as some kind of metadata, but for now I think this is the right
tradeoff.