Fix deletion errors
Batched updates to the rescue.

We had a problem because we have a dependency in the stores (expressed
via `waitFor` like this:

  NotesStore -> FilteredNotesStore -> NotesSelectionStore

As soon as NotesStore emitted its change event, we'd re-render. At this
point, we'd have deleted notes, and our selection could still be
pointing at a now-non-existent note.

I first "fixed" this by getting NotesStore and FilteredNotesStore to not
emit change events for deletion actions, knowing that listeners to those
stores also happened to listen to NotesSelectionStore (I additionally
had to tweak NoteList.react to have a single `_update` method that
refreshed all state from all stores entire).

Switching to React batched updates does that same trick for us with less
jank, and is probably a good idea for perf reasons anyway. I believe
this pattern may allow me to clean up some other bits of code too,
provided the best practice is followed of making your store listeners
very simple users of `setState` which don't do a lot of derivative
calculations (those should happen just-in-time for `render` anyway).

Another observation: for most components it is going to be a smell to
listen to NotesStore at all; almost all of them are going to want to
listen to FilteredNotesStore instead.