Entirely disable no-use-before-define lint
I don't like the restructuring that I've have had to do here in order to
pacify the linter (exposing a function which should be private within
the module closure), so let's just turn it off.

This isn't C, and we shouldn't have to declare methods in header files
before they used.

It's true that this lint could in theory catch some not-kosher
abuses/misues of variable hoisting, but I'm not too worried about those
cropping up all that often.