Silence a lint
This module is being imported for side-effects only,
which may itself be a smell, but I am not sure how else
to avoid that, given that nobody is going to be querying
the GitStore for anything, so there is no reason why it
would be otherwise required. Perhaps this is a hint that
it shouldn't be a store, but for now I'm going to run
with it.