Fix buggy SystemStore implementation
Leaving this in the history rather than amending it as a lesson: always
include a `catch` whenever you use promises. In theory Bluebird allows
you to be a bit sloppy with this, but in this case it didn't catch my
mistake (passing `...args` rather than an array to `Promise.all`).

While I'm fixing this, also tightening up the output parsing, converting
output like "255\n" to 255.