Remember and restore cursor positions on blur
Also, on note changes (because it is possible to change notes without
blurring via Command-{J,K}).

This is just a rough cut. I'll DRY it up when I have a bit more time. I
went with a lightweight object as map rather than a full-blown store for
tracking this state, as Flux is probably overkill for this.

(Will likely reconsider that when/if I decide to persist cursor
positions across restarts, which will require me having to dump them
somewhere persistent, and also guarantee stable note ids across
restarts, which I currently don't do.)