Add first draft of duplicate title handling
As noted inline, we can disambiguate among three "foo" notes by giving
them titles like "foo.txt", "foo.1.txt" and "foo.2.txt".

We have two basic use cases:

(1) Find out whether a particular path is in use (fast); this suggests
use of a map or a set.
(2) If a path is in use, find the next available disambiguating path;
this one should be rarer and therefore doesn't need to be fast, so I
went with the very simple structure of doing a linear probe of the map
looking for the next available slot.

There are still some unhandled edge cases to deal with (noted inline),
and some refactoring to make this cleaner. I am not super happy that the
management of the `pathsMap` ends up being sprinkled throughout the
`NotesStore`.