Add test to show failure edge case in Heap implementation
My spider sense was tingling when I wrote that last test, wondering if
alternating insertion and extraction would expose an edge case. Turns
out, it did, although the actual alternation has nothing to do with it.

If we insert jobs in this order:

- low-pri (1)
- high-pri
- low-pri (2)

This is actually happening under the covers:

- insert low-pri (1) in first empty slot (slot 0)
- insert high-pri in first empty slot (slot 1)
- bubble high-pri up until heap property is restored, effectively
  swapping the two items
- insert low-pri (2) in the first empty slot (slot 2)
- heap property is intact, so we do nothing

When it comes time to extract:

- extract first item, the high-pri job from slot 0
- move last item to root, ie. low-pri (2) moves from slot 2 to slot 0
- heap property is intact, so we do nothing
- extract next item, low-pri (2), violating FIFO expectation

I'll fix this in a separate commit.