Rename a Heap method for readability
This method is not as transparently named as it could be so change it;
it's original name was fruit of the fact that this heap implementation
started off as configurable to be either a max or a min heap.

Turns out that in looking at this I've spotted a bug in the
implementation; our comparison is not going to work with wrapped values,
and we're failing to use the key getter anyway. I'll fix that in a
separate commit.