Fix _pendingDeletion bug in OmniBar
Not strictly related to https://github.com/wincent/corpus/issues/17 but
discovered while working on it.

Give note "Foobar":

- Type "Foo" (completes to "Foobar" with "bar" selected)
- Hit backspace (shows "Foo")
- Hit backspace again (shows "Fo")
- Hit "o" (doesn't update!)

Another lesson in the pitfalls of mutable state[*]; this time the
lingering value of `_pendingDeletion` was the culprit.

[*]: Given the complexity of the interaction here between user-driven
events, multiple stores, and large parts of the UI, it's a necessary
evil for now[**], unfortunately.

[**]: Until I think of something better.