Horrid workaround for non-selected OmniBar text at startup
I've dug into this for over half an hour and the results are puzzling.

Adding CSS like this:

```
html *:focus {
  border: 1px solid #f00 !important;
  outline: 1px solid #f00 !important;
}
```

Shows that the OmniBar does indeed get and keep focus at startup
(`console.log` also confirms). Nevertheless, it is a kind of faux focus.
Our attempts to set the selection range don't have any visible effect,
unless they occur after a delay. Typing doesn't work as expected either;
the OmniBar doesn't update correctly.

I added logging in all the other places were we listen for `focus`
events, and none of them are getting the focus. Likewise `keydown`
events are clearly getting routed to the OmniBar. I give up; I believe
this used to work just fine, so something too subtle is going on here.

Adding listeners for "selectionchange", "select" and "selectstart" shows
that we're actually getting a few attempts to change the selection:

- one with range [0, 0] on first mount, which makes sense because we have
  no note title to work with at that point
- one with range [titleLength, titleLength], which corresponds to moving
  the cursor to the end of the title (not sure why this one is
  happening)
- one with range [0, titleLength]; this is the one I'm explicitly
  kicking off
- one with range [0, 0], which is what causes us to lose the selection,
  and I have no idea where it's coming from

In the end, I'm going to ship this horrible hacky workaround. The value
of 300 is empirically determined to apparently work. 200ms was too fast.
Why the heck this is remains a mystery.

Partially addresses: https://github.com/wincent/corpus/issues/19