Implement "Show in Finder"
The output on failure of the `open` command here sucks a bit (just
prints "1"), so I should look into that, although it's not a high-pri
because I don't expect this command to ever fail, really.

Also, as I previously noted, I need to build some better abstractions
for menu management. "Two or more, use a for", they say, and we're
already at three awkwardly managed menu items so far, so I'll have to
tackle this at some point soon.