Revert timeout hack
This affects _all_ typing of title-matching substrings, not just the
first focus event on launch (obviously), so this is an unacceptable
workaround for: https://github.com/wincent/corpus/issues/19

Back to the old drawing board. I could make the timeout apply
conditionally, or I could keep searching for the real cause of these
unwanted selection changes.