Handle lowercase keypresses in printableFromKeyEvent
The browser always sends an uppercase key identifier, regardless of
whether the shift key is pressed. Let's add some special casing to
faithfully return an upper or lowercase value as appropriate.

In practice, you might not have noticed a problem here, because of the
way we do title matching (the key you're pressing probably matches a
title, and when it does, we end up using whatever the matching
capitalization is from the title).

Another way I thought about doing this was with a simple:

```
return event.shiftKey ? string : string.toLowerCase();
```

But whatever. By the time I'd thought of it I'd already written and
tested the code in this commit.