Autosave when idle
No idea if this is too eager or not, but I'm going to save after 5
seconds of inactivity after any change. This will probably result in
quite a few Git commits being made, but I am paranoid about data loss,
so I'd prefer to start with an aggressive default like this.