Get rid of ugly linear scan in ContentEditable.react
When I originally wrote this, the linear scan was necessary. Since then,
the `FilteredNotesStore` learnt how to decorate the returned notes with
an "index" property containing the index of the note within the
`NotesStore`.

The notes that we get in `ContentEditable.react` (via `this.props.note`)
come from the `FilteredNotesStore`, so they have this "index" property.
Let's use it!