Remove inaccurate comment about transducers
This comment wasn't correct. Even with a transducer, you end up having
to execute the map function for every item in the collection. It might
be marginally faster, but it doesn't enable you to only do the mapping
on the subset of records that would be filtered, because you can't know
which ones would be filtered until you do the mapping.

Nevertheless, not a performance bottleneck right now, so forgetting
about transducers.