De-dupe NoteAnimationStore change events
I don't think we ever need to notify of multiple changes for the same
note.

If you're typing and generating multiple `NOTE_BUBBLE_STARTED` actions,
on the first one the store will update and fire a change event and the
animation will start. If you keep typing there is no need to change
anything; we can just let the animation continue.

As soon as you switch to another note and start typing we'll generate a
new action and a new "change" event.