Fix up note bubbling, a little
Still horribly broken, due to races noted inline. It _looks_ like it
works if you're just inspecting the UI, but if you look at the internals
you can see it's horribly wrong.

I think I'm going to refactor the stores a lot to liberate myself of
these race/ordering issues, so making this commit as a kind of
checkpoint, even though the app isn't usable right now like this.