Minor clean up of note animations
Some better comments, fix a lint, and remove an unnecessary check.

I understand the nature of the problem a bit better now. By hard coding
_all_ notes to animate in a certain direction, I can see that the notes
that get bumped down in the list remain still; the note that got removed
and reinserted at the top, and everything below its original position,
all animate successfully.

So this looks like a React issue, which I'll have to figure out how to
work around. I believe my `render()` method is returning the right
inline styles, but it is possible that when we rejig the DOM, we end up
removing the original nodes and inserting new nodes, so the browser
thinks there is no initial state to translate from, and it just jumps to
the final state.