Fix note bubbling
So this is pretty complicated and gnarly, and I'd like to make it
clearer, but it appears to work.

We work around our React DOM manipulation woes by doing a couple of
separate render passes, as explained in the comments. Effectively state
transitions as follows:

(0) (Initial state) `bubbling` is `null`, and `animating` is `false`.
(1) Note gets moved within the `NoteStore` to the top of the list; we
set state: `bubbling` is the original index of the bubbled note (this
triggers a re-render, so we offset the moved notes to look as though
they were still in their old position).
(2) Once that render is complete, we set state again: `bubbling` is
still the original index of the bubbled note, but we now set `animating`
to `true` (this triggers re-render, where the animation actually
starts).
(3) The transition ends, so we re-set state to `bubbling: null,
animating: false`; this triggers yet another render, but nothing
actually happens.

I might be able to reduce the number of states, renders, or interactions
between them here, but I do know that I need at least two renders.