Reduce number of re-renders for note bubbling
Prior to this commit, we effectively had four states for a bubble
animation:

  bubbling | animating | comment
  -----------------------------------------------------------------
  null     | false     | initial state, first render
  N        | false     | render at offset, second render
  N        | true      | render with animation, third render
  null     | false     | back to initial state again, fourth render

That last render, however, is just needless churn; we wind up removing
some inline styles from the DOM (specifically, the `transition`
property), but they have no visual effect because the transition has
already completed at this point. I haven't checked this in the dev
tools, but I wouldn't be surprised if this triggered something expensive
(perhaps not a layout, but I'd expect a style recalc and possibly a
repaint).

With this, we now just do:

  bubbling | animating | comment
  -----------------------------------------------------------------
  null    | false      | initial state, first render
  N       | false      | render at offset, second render
  N       | true       | render with animation, third render

Once the animation completes, we don't bother re-setting state, because
when it comes time to perform another animation we just tack on another
two cycles:

  bubbling | animating | comment
  -----------------------------------------------------------------
  N       | false      | render at offset, first render
  N       | true       | render with animation, second render