Don't swallow first input on note bubbling
First edit was getting dropped on the floor here because of the
following sequence of events:

- `_onChange`, we fire `Actions.noteBubbled`; this mutates the note
  instance (by setting a new `index` value on it)
- `setState` to the new content; no render yet due to batched rendering
  (this is an event handler)
- `componentWillReceiveProps` fires, and we think the note is changing
  because it is no longer `===`
- we `setState` to the "new" text from the new note, but that's actually
  the old text because we never mutated the text of the note, just its
  index

The solution then, is to be more careful to check for changes in
note-identity rather than contents, because _those_ get taken care of in
`_onChange`.

I also toyed with actually mutating the `text` of the note, but didn't
seem like a clear win. Likewise, note that I'm not updating the mtime of
the note; I think I prefer that to match the filesystem mtime.