Fix unwanted OmniBar focusing
Fixes the first bug mentioned in:

https://github.com/wincent/corpus/issues/31

Testing, this doesn't seem to have broken anything else. Except for one
small thing: if the Note view itself has focus, we no longer focus the
OmniBar (and might loop back and fix that, although checking nvALT right
now, it doesn't seem to have that behavior anyway).

It's almost too painful to explain the code path that was triggering
this, but effectively, when the search in progress ended up changing the
selected note, we'd fire an `omniBarFocused` action, which would wind up
in `_updateFocus` in the `OmniBar` component, selecting everything.