Start using async/await
The future is now. Starting off in a small way because some of the
gnarlier promise-based stuff will be a bit hard to convert. Is this
better? I don't know; is:

```
try {
  const data = await readFile(configFile);
  config = config.mergeWith(merger, JSON.parse(data.toString()));
} catch(error) {
  warn(`Reading ${configFile}: ${error.message}`);
}
Actions.configLoaded();
```

better than?

```
readFile(configFile)
  .then(data => (
    config = config.mergeWith(merger, JSON.parse(data.toString()))
  ))
  .catch(error => warn(`Reading ${configFile}: ${error.message}`))
  .finally(Actions.configLoaded);
```

It might be, but for the fact that it requires us to wrap our code in an
`async` function (because the constructor itself can't be async); and
this is going to happen everywhere... as soon as you want one leaf
function to be async, everything else that calls it all the way to the
top has to be async too. This is why I am not sure I'm going to love
this, but this whole app is a playground anyway, so I may as well
experiment.