Switch OperationsQueue clients to async/await
This is a big change, and only lightly tested, so beware.

I did a before-and-after benchmark comparing the start time of the old
implementation of `loadNotes` (which used Bluebird's `Promise.map` with
a concurrency of `5`) with the new implementation (which uses `async*`,
the equivalent of `Promise.all`, which means we load the first 20 notes
in parallel and then all the rest in parallel). The average of 5 runs
was:

- before: 325ms
- after: 427ms

for my current corpus of just under 600 notes. This is probably good
enough, but I can look at batching it later on if I really want to
optimize start time.