Disable `global-strict` ESLint rule
According to the ESLint documentation, this rule is deprecated and
superseded by the `strict` rule, and it will be removed in ESLint 1.0
[0]. However, in the meantime, it was blowing up all over the place with
errors like:

> error  Use the function form of "use strict"  global-strict

The `strict` rule has already been enabled and configured to prefer the
global version of "use strict", so we should be covered there.

I considered disabling this by using the simpler `0` instead of `[0]`,
but I noticed that other rules were already using `[0]` here, so I went
with consistency.

[0]: http://eslint.org/docs/rules/global-strict